created separate component for each
